split FilterPanel to several new components, delete characters folder, and move its components to upper folder
