+++ src/components/Cards.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Card from './Card';
 import { Box, Typography } from '@mui/material';
-import { useCharactersContext } from '../../context/CharactersContext';
+import { useCharactersContext } from '../context/CharactersContext';
 
 const Cards = () => {
   const { filteredCharacters } = useCharactersContext();
