+++ src/components/FilterPanel/index.js
@@ -1,3 +1,15 @@
 import FilterPanel from './FilterPanel';
+import FilterModal from './FilterModal';
+import FilterContent from './FilterContent';
+import FilterNameField from './FilterNameField';
+import GenderRadioButtons from './GenderRadioButtons';
+import StatusRadioButtons from './StatusRadioButtons';
 
-export default FilterPanel;
+export {
+  FilterPanel,
+  FilterModal,
+  FilterContent,
+  FilterNameField,
+  GenderRadioButtons,
+  StatusRadioButtons,
+};
